requirements.txt
pytest==8.0.0
requests==2.31.0
python-dotenv==1.0.0
pytest-html==4.1.1  # Generate HTML test reports
pytest-xdist==3.5.0  # Run tests in parallel
